Finished CRUD task routes

+++ app/routes.py
@@ -1,8 +1,9 @@
-
 from os import abort
 from app import db
 from app.models.task import Task
 from flask import Blueprint, jsonify, abort, make_response, request
+from datetime import datetime
+
 
 tasks_bp = Blueprint("tasks_bp", __name__, url_prefix="/tasks")
 
@@ -12,7 +13,7 @@ def validate_task(task_id):
     except:
         abort(make_response({"message":f"Task {task_id} invalid"}, 400))
 
-    task = task.query.get(task_id)
+    task = Task.query.get(task_id)
 
     if not task:
         abort(make_response({"message":f"Task {task_id} not found"}, 404))
@@ -23,16 +24,16 @@ def validate_task(task_id):
 @tasks_bp.route("", methods=["POST"])
 def create_task():
     request_body = request.get_json()
-    if not 'description' in request_body or not 'title' in request_body:
-        return make_response(jsonify({"details": "Invalid data"})), 400
-    new_task = Task(title=request_body["title"],
-                    description=request_body["description"],
-                    completed_at=request_body["completed_at"])
-
+    if ("title" not in request_body) or ("description" not in request_body):
+        return make_response({"details": "Invalid data"}, 400)
+    new_task = Task(title = request_body["title"],
+                    description = request_body["description"],
+                    completed_at = request_body["completed_at"] if "completed_at" in request_body else None)
+        
     db.session.add(new_task)
     db.session.commit()
 
-    return make_response(jsonify(f"Task {new_task.title} successfully created"), 201)
+    return make_response({"task": new_task.response_dict()}, 201)
 
 @tasks_bp.route("", methods=["GET"])
 def read_all_tasks():
@@ -43,47 +44,36 @@ def read_all_tasks():
     else:
         tasks = Task.query.all()
 
-    tasks_response = []
-    for task in tasks:
-        tasks_response.append(
-            {
-                "id": task.id,
-                "title": task.title,
-                "description": task.description,
-                "is_complete": task.completed_at
-            }
-        )
+    tasks_response = [task.response_dict() for task in tasks]
+
     return jsonify(tasks_response)
 
 @tasks_bp.route("/<task_id>", methods=["GET"])
-def read_one_task(task_id):
+def get_task(task_id):
     task = validate_task(task_id)
-    return  {
-                "id": task.id,
-                "title": task.title,
-                "description": task.description,
-                "is_complete": task.completed_at
-            }
+    return make_response({"task": task.response_dict()}, 200)
 
 @tasks_bp.route("/<task_id>", methods=["PUT"])
 def update_task(task_id):
     task = validate_task(task_id)
+    if task is None:
+        return make_response({'message': f'Task {task.id} not found'}, 404)
 
-    request_body = request.get_json()
-
-    task.title = request_body["title"]
-    task.description = request_body["description"]
-    task.completed_at = request_body["completed_at"]
+    form_data = request.get_json()
+    task.title = form_data["title"]
+    task.description = form_data["description"]
 
     db.session.commit()
-
-    return make_response(jsonify(f"Task #{task.id} successfully updated"))
+    return make_response({"task": task.response_dict()}, 200)
 
 @tasks_bp.route("/<task_id>", methods=["DELETE"])
 def delete_task(task_id):
+    task = Task.query.get(task_id)
+    if task is None:
+        return make_response("", 404)
     task = validate_task(task_id)
 
     db.session.delete(task)
     db.session.commit()
 
-    return make_response(jsonify(f"Task 1 {task.title} successfully deleted"))
+    return make_response({'details' : f'Task {task.id} "{task.title}" successfully deleted'})
